+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 09/06/2024
+
 ### Added
 
 - `MaybeUninit[T]` type.
@@ -15,16 +17,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Vec.push_within_capacity`
 - `Vec.reserve`
 - `Some.take_if`
+- `iter.repeat`
 
 ### Changed
 
 - `Option.take` now return `Option[T]` instead of `None`
 - Stabilized `Vec.with_capacity` and its fellow methods.
 - Calling `RefCell.copy` now increment its ref count when called.
-- The repr of `Some[None]` is now `None` instead of `Some(None)`
+- The repr of `Some(None)` is now `None` instead of `Some(None)`
 - Passing a `list` variable when constructing a `Vec` now will point to that list instead of copying it.
 - Better documentation on `Vec`
 - `vec.vec` is renamed to `vec.from_args`
+- `iter.empty` now returns `Empty[T]` instead of `Iter[Never]` with better type hinting.
+- Better `repr` for iterators.
+- Changed the default value of `Some` from `NoneType` to `Some(None)`
 
 ## 0.0.6 - 05/04/2024